flutter/flutter_powersync_integration/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
cryptography==43.0.0
